Refactor API calls into a separate NetworkService class, using an event bus for communication.
This has several important advantages:

 - Activities / Fragments no longer have to deal with the Retrofit-generated API
directly. This improves code clarity and ensures that the view is truly a view.

 - Using an event bus decouples the view from the network, so a LoadPostsEvent
could be satisfied by a mere database query instead of a full-blown network
request, behind-the-scenes. This lays the foundation for offline mode.

 - Clean code: woe to BaseActivity.sAuthToken and the Globals class!

 - Overall the code now probably lends itself well to automated testing, which
is always a good sign, irrespective of whether we actually use tests or not.