Start new drafts directly in Edit tab + fix an orientation change crash.
There was a consistent crash that could be reproduced with these steps:

1. Create a new draft, a dummy post with id=0 is passed to
   PostViewActivity through its Intent extras.
2. In the background, the post is uploaded and a PostReplacedEvent is
   fired, which also updates PostViewActivity's mPost.
3. Rotate the screen.
4. Delete the draft => crash!!!

This happened because after (3) the Activity was re-created with the
original Intent extras, so the replaced post was lost. The fix was to
save the latest values corresponding to the Intent extras in
onSavedInstanceState, and using that during Activity re-creation.