Fix #145: Users with only "Author" role are not handled correctly.
Here's how role-based permissions work as of Ghost v0.11.0:

There are 4 roles: Owner, Admin, Editor, Author. Each of these roles is
effectively a proper subset of the one following it, so an Author has
fewer permissions than an Editor, and so on.

Coming to permissions for viewing/editing posts: Owners, Admins and
Editors can view and edit ALL posts by ALL users. So Editors can even
view and edit an Owner's posts, which is a little counter-intuitive. The
only special case is Authors, who can only view and edit their *own*
posts.