Workaround to avoid image upload triggering spurious post conflict.
This prevents the auto-save in onPause() when an image upload is in
progress (the file picker opens a new activity unfortunately).

There is no better way to handle this currently, because the PostReplacedEvent
arrives after the activity is already gone, so the activity's Post is
out-of-sync and hence deemed a "conflict" on the next save.

This should significantly reduce the number of post conflict events seen on
Crashlytics.