Resurrect image upload code removed in commit b0c8165, because a URI cannot always be converted to a filesystem path.
As explained here - http://stackoverflow.com/a/39490576/504611 - a URI from
ACTION_PICK cannot reliably be converted to a URI. It is easy to observe this
when, for example, picking a file via Clean File Manager or a Google Photos pic
that is not already on the device.

The hybrid approach should be more reliable while preserving data integrity as
far as possible - this is not perfect but better than the alternatives.

References #155.